refactor: extract GEMINI/LLM into feature/llm (part of #778)
Move the GEMINI statement family out of core internal/mycli into
internal/mycli/feature/llm through the Feature seam: the statement, the
two-phase compose flow, the tool-use Developer Knowledge client, the doc
cache (session-scoped, now a FeatureState holder closed via io.Closer),
and the embedded doc catalog all move; core no longer imports
google.golang.org/genai, genaischema, developerknowledge-go, or
spanner-docs-embed.

First use of the Flags/KongVars/ApplyFlags seam: the --vertexai-* flags
move to Feature.Flags (kong.Plugins), the default-model/location help
values to Feature.KongVars, and flag application routes through the
registry via applyFeatureFlags in initializeSystemVariables - after the
registry is available and before --set processing, preserving the
pre-extraction precedence (flag < --set; unset pointer flags do not
clobber TOML/--set). The CLI_VERTEXAI_* variables move to FeatureVars
with verbatim descriptions and the same defaults; docs/system_variables.md
is byte-identical. kong.Plugins renders plugin flags at the end of
--help, so the three --vertexai-* lines move to the end of the flag list
in the regenerated README readme-help block - the flag text itself is
byte-identical, and in the slim variant these lines disappear entirely,
so optional-feature flags listing last mirrors the statement-table
layout. GeminiStatement keeps implementing no marker interfaces (not
detached-compatible, not READONLY-classified), unchanged from core.

Core test adjustments: the TOML nested-typo case now uses the [emulator]
prefix (vertexai left core; the table name must be a non-flag prefix of
real flags for the unknown-key rejection path); CLI_VERTEXAI_* Set/Get
and enumeration/default coverage moves to the external llm_variables_test
with the real llm feature registered.
