+++ internal/mycli/app.go
@@ -40,11 +40,12 @@ import (
 )
 
 const (
-	defaultPrompt           = "spanner%t> "
-	defaultPrompt2          = "%P%R> "
-	defaultVertexAIModel    = "gemini-3-flash-preview"
-	defaultVertexAILocation = "global"
-	DefaultAnalyzeColumns   = "Rows:{{.Rows.Total}},Exec.:{{.ExecutionSummary.NumExecutions}},Total Latency:{{.Latency}}"
+	defaultPrompt  = "spanner%t> "
+	defaultPrompt2 = "%P%R> "
+	// defaultVertexAIModel / defaultVertexAILocation moved to
+	// internal/mycli/feature/llm (#778); the GEMINI feature supplies them to the
+	// kong parser as help-template vars in the full variant.
+	DefaultAnalyzeColumns = "Rows:{{.Rows.Total}},Exec.:{{.ExecutionSummary.NumExecutions}},Total Latency:{{.Latency}}"
 )
 
 var DefaultParsedAnalyzeColumns = lo.Must(customListToTableRenderDefs(DefaultAnalyzeColumns))
