+++ internal/mycli/main_test.go
@@ -89,8 +89,6 @@ func Test_initializeSystemVariables(t *testing.T) {
 				QueryMode:                 ptrCaseInsensitiveEnumValue("PLAN"),
 				Strong:                    true,
 				ReadTimestamp:             "", // Strong takes precedence
-				VertexAIProject:           "vertex-project",
-				VertexAIModel:             lo.ToPtr("gemini-1.0-pro"),
 				DatabaseDialect:           ptrCaseInsensitiveEnumValue("POSTGRESQL"),
 				ImpersonateServiceAccount: "test-sa@example.com",
 				EnablePartitionedDML:      true,
@@ -148,12 +146,9 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReadOnly:          true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelInfo,
-					VertexAIProject:  "vertex-project",
-					VertexAIModel:    "gemini-1.0-pro",
-					VertexAILocation: defaultVertexAILocation,
-					DatabaseDialect:  databasepb.DatabaseDialect_POSTGRESQL,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:        slog.LevelInfo,
+					DatabaseDialect: databasepb.DatabaseDialect_POSTGRESQL,
+					FuzzyFinderKey:  "C_T",
 				},
 				Internal: InternalVars{
 					ProtoDescriptorFile: []string{"testdata/protos/singer.proto"},
@@ -204,10 +199,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 				Params: make(map[string]ast.Node),
 			},
@@ -259,10 +252,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 				Params: make(map[string]ast.Node),
 			},
@@ -299,10 +290,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -338,10 +327,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -377,10 +364,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -435,10 +420,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 				Params: make(map[string]ast.Node),
 			},
@@ -501,10 +484,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -544,10 +525,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -588,10 +567,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -631,10 +608,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -675,10 +650,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -714,10 +687,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -751,10 +722,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -790,10 +759,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -829,10 +796,8 @@ func Test_initializeSystemVariables(t *testing.T) {
 					ReturnCommitStats: true,
 				},
 				Feature: FeatureVars{
-					LogLevel:         slog.LevelWarn,
-					VertexAIModel:    defaultVertexAIModel,
-					VertexAILocation: defaultVertexAILocation,
-					FuzzyFinderKey:   "C_T",
+					LogLevel:       slog.LevelWarn,
+					FuzzyFinderKey: "C_T",
 				},
 			},
 			wantErr: false,
@@ -945,10 +910,8 @@ func Test_newSystemVariablesWithDefaults(t *testing.T) {
 			ReturnCommitStats: true,
 		},
 		Feature: FeatureVars{
-			LogLevel:         slog.LevelWarn,
-			VertexAIModel:    defaultVertexAIModel,
-			VertexAILocation: defaultVertexAILocation,
-			FuzzyFinderKey:   "C_T",
+			LogLevel:       slog.LevelWarn,
+			FuzzyFinderKey: "C_T",
 		},
 		Params: make(map[string]ast.Node),
 	}
